Refactor dashboard model into modular builders
