+++ src/ui/dashboard/model.js
@@ -1,194 +1,16 @@
 import { formatHours, formatMoney } from '../../core/helpers.js';
-import { getAssetState } from '../../core/state.js';
-import { getAssets, getHustles, getUpgrades } from '../../game/registryService.js';
-import { getNicheRoster, getNicheWatchlist } from '../../game/assets/niches.js';
+import { clampNumber, clampScore, formatPercent, describeDelta } from './formatters.js';
+import { buildDailySummaries } from './passiveIncome.js';
 import {
-  canPerformQualityAction,
-  getNextQualityLevel,
-  getQualityActions,
-  getQualityTracks,
-  performQualityAction
-} from '../../game/assets/quality.js';
-import { instanceLabel } from '../../game/assets/helpers.js';
-import { KNOWLEDGE_TRACKS, getKnowledgeProgress } from '../../game/requirements.js';
-
-function clampNumber(value) {
-  const numeric = Number(value);
-  return Number.isFinite(numeric) ? numeric : 0;
-}
-
-function clampScore(value) {
-  if (!Number.isFinite(value)) return null;
-  return Math.max(0, Math.min(100, Math.round(value)));
-}
-
-function getMetricLabel(entry = {}) {
-  return entry.label
-    || entry?.definition?.label
-    || entry?.definition?.name
-    || entry.key
-    || 'Metric';
-}
-
-function extractIconPrefix(label) {
-  if (typeof label !== 'string') return '';
-  const match = label.match(/^([^\w]*)/u);
-  return match ? match[1].trim() : '';
-}
-
-function formatTimeEntries(summary = {}) {
-  const entries = Array.isArray(summary.timeBreakdown) ? summary.timeBreakdown : [];
-  return entries.map(entry => {
-    const hours = clampNumber(entry?.hours);
-    return {
-      key: entry?.key,
-      label: getMetricLabel(entry),
-      value: `${formatHours(hours)} today`,
-      hours,
-      category: entry?.category || entry?.definition?.category || 'general',
-      definition: entry?.definition
-    };
-  });
-}
-
-function formatPayoutEntries(entries = []) {
-  return entries.map(entry => {
-    const amount = clampNumber(entry?.amount);
-    const baseLabel = getMetricLabel(entry);
-    let label = baseLabel;
-
-    if (entry?.source?.type === 'asset') {
-      const prefix = extractIconPrefix(entry?.label);
-      const name = entry?.source?.name || baseLabel;
-      const decorated = entry?.source?.count > 1 ? `${name} (${entry.source.count})` : name;
-      label = prefix ? `${prefix} ${decorated}`.trim() : decorated;
-    }
-
-    return {
-      key: entry?.key,
-      label,
-      value: `$${formatMoney(amount)} today`,
-      amount,
-      category: entry?.category || entry?.definition?.category || 'general',
-      definition: entry?.definition,
-      stream: entry?.stream,
-      source: entry?.source
-    };
-  });
-}
-
-function formatSpendEntries(entries = []) {
-  return entries.map(entry => {
-    const amount = clampNumber(entry?.amount);
-    return {
-      key: entry?.key,
-      label: getMetricLabel(entry),
-      value: `$${formatMoney(amount)} today`,
-      amount,
-      category: entry?.category || entry?.definition?.category || 'general',
-      definition: entry?.definition
-    };
-  });
-}
-
-function formatStudyEntries(entries = []) {
-  return entries.map(entry => {
-    const hours = clampNumber(entry?.hoursPerDay);
-    const remaining = Math.max(0, clampNumber(entry?.remainingDays));
-    const status = entry?.status || (entry?.studiedToday ? 'scheduled' : 'waiting');
-    return {
-      trackId: entry?.trackId,
-      label: `📘 ${entry?.name}`,
-      value: `${formatHours(hours)} / day • ${remaining} day${remaining === 1 ? '' : 's'} left (${status})`,
-      hoursPerDay: hours,
-      remainingDays: remaining,
-      studiedToday: Boolean(entry?.studiedToday),
-      status
-    };
-  });
-}
-
-function buildSummaryPresentations(summary = {}) {
-  return {
-    timeEntries: formatTimeEntries(summary),
-    earningsEntries: formatPayoutEntries(Array.isArray(summary.earningsBreakdown) ? summary.earningsBreakdown : []),
-    passiveEntries: formatPayoutEntries(Array.isArray(summary.passiveBreakdown) ? summary.passiveBreakdown : []),
-    spendEntries: formatSpendEntries(Array.isArray(summary.spendBreakdown) ? summary.spendBreakdown : []),
-    studyEntries: formatStudyEntries(Array.isArray(summary.studyBreakdown) ? summary.studyBreakdown : [])
-  };
-}
-
-function getQualitySnapshot(instance = {}) {
-  const level = Math.max(0, clampNumber(instance?.quality?.level));
-  const progress = instance?.quality?.progress && typeof instance.quality.progress === 'object'
-    ? { ...instance.quality.progress }
-    : {};
-  return {
-    level,
-    progress
-  };
-}
-
-function resolveProgressAmount(action, context) {
-  if (!action) return 0;
-  if (typeof action.progressAmount === 'function') {
-    try {
-      const amount = Number(action.progressAmount(context));
-      if (Number.isFinite(amount) && amount > 0) {
-        return amount;
-      }
-    } catch (error) {
-      return 0;
-    }
-  }
-  if (Number.isFinite(Number(action.progressAmount))) {
-    const numeric = Number(action.progressAmount);
-    return numeric > 0 ? numeric : 0;
-  }
-  if (action.progressKey) {
-    return 1;
-  }
-  return 0;
-}
-
-function getProgressPerRun(asset, instance, action, state) {
-  if (!asset || !instance || !action) return 0;
-  const quality = getQualitySnapshot(instance);
-  const context = {
-    state,
-    definition: asset,
-    instance,
-    quality,
-    upgrade: id => state?.upgrades?.[id]
-  };
-  return resolveProgressAmount(action, context);
-}
-
-function estimateRemainingRuns(asset, instance, action, remaining, state) {
-  if (!Number.isFinite(remaining) || remaining <= 0) {
-    return 0;
-  }
-  const progressPerRun = getProgressPerRun(asset, instance, action, state);
-  if (!Number.isFinite(progressPerRun) || progressPerRun <= 0) {
-    return null;
-  }
-  return Math.max(1, Math.ceil(remaining / progressPerRun));
-}
-
-function describeDelta(popularity = {}) {
-  const raw = Number(popularity.delta);
-  if (!Number.isFinite(raw)) return 'Fresh reading';
-  if (raw === 0) return 'Holding steady';
-  const sign = raw > 0 ? '+' : '';
-  return `${sign}${raw} vs yesterday`;
-}
-
-function formatPercent(value) {
-  if (!Number.isFinite(value)) return '0%';
-  const percent = Math.round(value * 100);
-  const sign = percent > 0 ? '+' : '';
-  return `${sign}${percent}%`;
-}
+  buildQuickActionModel,
+  buildAssetActionModel,
+  buildQuickActions,
+  buildAssetUpgradeRecommendations
+} from './quickActions.js';
+import { buildStudyEnrollmentActionModel } from './knowledge.js';
+import { getAssetState } from '../../core/state.js';
+import { getAssets, getUpgrades } from '../../game/registryService.js';
+import { getNicheRoster, getNicheWatchlist } from '../../game/assets/niches.js';
 
 function describeTrendStatus(entry) {
   if (!entry) return 'Steady';
@@ -351,231 +173,6 @@ export function buildNicheViewModel(state) {
   };
 }
 
-function computeTimeCap(state = {}) {
-  const base = clampNumber(state.baseTime);
-  const bonus = clampNumber(state.bonusTime);
-  const daily = clampNumber(state.dailyBonusTime);
-  return Math.max(0, base + bonus + daily);
-}
-
-function describeQueue(summary = {}) {
-  const { timeEntries } = buildSummaryPresentations(summary);
-  const entries = Array.isArray(timeEntries) ? timeEntries : [];
-  if (!entries.length) {
-    return [
-      {
-        id: 'queue:empty',
-        label: 'Nothing queued yet',
-        detail: 'Open hustles to schedule your next move.',
-        hours: 0,
-        state: 'idle'
-      }
-    ];
-  }
-
-  return entries.slice(0, 6).map(entry => ({
-    id: entry.key || entry.label || 'queue:item',
-    label: entry.label,
-    detail: entry.value,
-    hours: clampNumber(entry.hours),
-    state: entry.category === 'maintenance' ? 'maintenance' : 'active'
-  }));
-}
-
-export function buildQuickActions(state) {
-  const items = [];
-  for (const hustle of getHustles()) {
-    if (hustle?.tag?.type === 'study') continue;
-    if (!hustle?.action?.onClick) continue;
-    const usage = typeof hustle.getDailyUsage === 'function' ? hustle.getDailyUsage(state) : null;
-    const remainingRuns = Number.isFinite(usage?.remaining)
-      ? Math.max(0, usage.remaining)
-      : Infinity;
-    const usageLimit = usage?.limit;
-    const repeatable = remainingRuns > 0 && (!Number.isFinite(usageLimit) || usageLimit !== 1);
-    const disabled = typeof hustle.action.disabled === 'function'
-      ? hustle.action.disabled(state)
-      : Boolean(hustle.action.disabled);
-    if (disabled) continue;
-    const payout = clampNumber(hustle.payout?.amount);
-    const time = clampNumber(hustle.time || hustle.action?.timeCost) || 1;
-    const roi = time > 0 ? payout / time : payout;
-    const payoutText = `$${formatMoney(payout)}`;
-    const timeText = formatHours(time);
-    items.push({
-      id: hustle.id,
-      label: hustle.name,
-      primaryLabel: typeof hustle.action.label === 'function'
-        ? hustle.action.label(state)
-        : hustle.action.label || 'Queue',
-      description: `${formatMoney(payout)} payout • ${formatHours(time)}`,
-      onClick: hustle.action.onClick,
-      roi,
-      timeCost: time,
-      payout,
-      payoutText,
-      durationHours: time,
-      durationText: timeText,
-      meta: `${payoutText} • ${timeText}`,
-      repeatable,
-      remainingRuns
-    });
-  }
-
-  items.sort((a, b) => b.roi - a.roi);
-  return items;
-}
-
-export function buildAssetUpgradeRecommendations(state) {
-  if (!state) return [];
-
-  const suggestions = [];
-
-  for (const asset of getAssets()) {
-    const qualityActions = getQualityActions(asset);
-    if (!qualityActions.length) continue;
-
-    const assetState = getAssetState(asset.id, state);
-    const instances = Array.isArray(assetState?.instances) ? assetState.instances : [];
-    if (!instances.length) continue;
-
-    const tracks = getQualityTracks(asset);
-
-    instances.forEach(instance => {
-      if (instance?.status !== 'active') return;
-
-      const quality = instance.quality || {};
-      const level = Math.max(0, clampNumber(quality.level));
-      const nextLevel = getNextQualityLevel(asset, level);
-      if (!nextLevel?.requirements) return;
-
-      const progress = quality.progress || {};
-      const requirements = Object.entries(nextLevel.requirements);
-      if (!requirements.length) return;
-
-      const assetIndex = assetState.instances.indexOf(instance);
-      const label = instanceLabel(asset, assetIndex >= 0 ? assetIndex : 0);
-      const performance = Math.max(0, clampNumber(instance.lastIncome));
-
-      requirements.forEach(([key, targetValue]) => {
-        const target = Math.max(0, clampNumber(targetValue));
-        if (target <= 0) return;
-        const current = Math.max(0, clampNumber(progress?.[key]));
-        const remaining = Math.max(0, target - current);
-        if (remaining <= 0) return;
-
-        const action = qualityActions.find(entry => entry.progressKey === key);
-        if (!action) return;
-        if (!canPerformQualityAction(asset, instance, action, state)) return;
-
-        const completion = target > 0 ? Math.min(1, current / target) : 1;
-        const percentComplete = Math.max(0, Math.min(100, Math.round(completion * 100)));
-        const percentRemaining = Math.max(0, 100 - percentComplete);
-        const track = tracks?.[key] || {};
-        const requirementLabel = track.shortLabel || track.label || key;
-        const timeCost = Math.max(0, clampNumber(action.time));
-        const moneyCost = Math.max(0, clampNumber(action.cost));
-        const effortParts = [];
-        if (timeCost > 0) {
-          effortParts.push(`${formatHours(timeCost)} focus`);
-        }
-        if (moneyCost > 0) {
-          effortParts.push(`$${formatMoney(moneyCost)}`);
-        }
-        const progressNote = `${Math.min(current, target)}/${target} logged (${percentComplete}% complete)`;
-        const meta = effortParts.length ? `${progressNote} • ${effortParts.join(' • ')}` : progressNote;
-        const actionLabel = typeof action.label === 'function'
-          ? action.label({ definition: asset, instance, state })
-          : action.label;
-        const buttonLabel = actionLabel || 'Boost Quality';
-        const remainingRuns = estimateRemainingRuns(asset, instance, action, remaining, state);
-        const repeatable = remainingRuns == null ? true : remainingRuns > 1;
-
-        suggestions.push({
-          id: `asset-upgrade:${asset.id}:${instance.id}:${action.id}:${key}`,
-          title: `${label} · ${buttonLabel}`,
-          subtitle: `${remaining} ${requirementLabel} to go for Quality ${nextLevel.level} (${percentRemaining}% to go).`,
-          meta,
-          buttonLabel,
-          onClick: () => performQualityAction(asset.id, instance.id, action.id),
-          performance,
-          completion,
-          remaining,
-          level,
-          timeCost,
-          remainingRuns,
-          repeatable
-        });
-      });
-    });
-  }
-
-  suggestions.sort((a, b) => {
-    if (a.performance !== b.performance) {
-      return a.performance - b.performance;
-    }
-    if (a.level !== b.level) {
-      return a.level - b.level;
-    }
-    if (a.completion !== b.completion) {
-      return a.completion - b.completion;
-    }
-    if (a.remaining !== b.remaining) {
-      return b.remaining - a.remaining;
-    }
-    return a.title.localeCompare(b.title);
-  });
-
-  return suggestions;
-}
-
-function computeAssetMetrics(state = {}) {
-  let activeAssets = 0;
-  let upkeepDue = 0;
-
-  for (const asset of getAssets()) {
-    const assetState = getAssetState(asset.id, state);
-    const instances = Array.isArray(assetState?.instances) ? assetState.instances : [];
-    instances.forEach(instance => {
-      if (instance?.status === 'active') {
-        activeAssets += 1;
-        if (!instance.maintenanceFundedToday) {
-          upkeepDue += clampNumber(asset.maintenance?.cost);
-        }
-      }
-    });
-  }
-
-  return { activeAssets, upkeepDue };
-}
-
-function computeStudyProgress(state = {}) {
-  const tracks = Object.values(KNOWLEDGE_TRACKS);
-  if (!tracks.length) {
-    return { percent: 0, summary: 'No study tracks unlocked yet.' };
-  }
-  let enrolled = 0;
-  let completed = 0;
-  let totalProgress = 0;
-  tracks.forEach(track => {
-    const progress = getKnowledgeProgress(track.id, state);
-    if (!progress.enrolled) return;
-    enrolled += 1;
-    if (progress.completed) {
-      completed += 1;
-      totalProgress += 1;
-    } else {
-      const fraction = Math.min(1, (progress.daysCompleted || 0) / Math.max(1, track.days));
-      totalProgress += fraction;
-    }
-  });
-  const percent = enrolled > 0 ? Math.round((totalProgress / Math.max(1, enrolled)) * 100) : 0;
-  const summary = enrolled
-    ? `${completed}/${enrolled} finished • ${percent}% average progress`
-    : 'No active study tracks.';
-  return { percent, summary };
-}
-
 function buildNotifications(state = {}) {
   const notifications = [];
 
@@ -650,329 +247,6 @@ function buildEventLog(state = {}) {
     .filter(Boolean);
 }
 
-function buildDailyStats(summary = {}) {
-  const presentations = buildSummaryPresentations(summary);
-  const totalTime = Math.max(0, clampNumber(summary.totalTime));
-  const setupHours = Math.max(0, clampNumber(summary.setupHours));
-  const maintenanceHours = Math.max(0, clampNumber(summary.maintenanceHours));
-  const otherTimeHours = Math.max(0, clampNumber(summary.otherTimeHours));
-  const timeSummaryText = totalTime > 0
-    ? `${formatHours(totalTime)} invested • ${formatHours(setupHours)} setup • ${formatHours(maintenanceHours)} upkeep • ${formatHours(otherTimeHours)} flex`
-    : 'No hours logged yet. Queue a hustle to get moving.';
-
-  const totalEarnings = Math.max(0, clampNumber(summary.totalEarnings));
-  const activeEarnings = Math.max(0, clampNumber(summary.activeEarnings));
-  const passiveEarnings = Math.max(0, clampNumber(summary.passiveEarnings));
-  const earningsSummaryText = totalEarnings > 0
-    ? `$${formatMoney(totalEarnings)} earned • $${formatMoney(activeEarnings)} active • $${formatMoney(passiveEarnings)} passive`
-    : 'Payouts will appear once you start closing deals.';
-
-  const totalSpend = Math.max(0, clampNumber(summary.totalSpend));
-  const upkeepSpend = Math.max(0, clampNumber(summary.upkeepSpend));
-  const investmentSpend = Math.max(0, clampNumber(summary.investmentSpend));
-  const spendSummaryText = totalSpend > 0
-    ? `$${formatMoney(totalSpend)} spent • $${formatMoney(upkeepSpend)} upkeep • $${formatMoney(investmentSpend)} investments`
-    : 'Outflows land here when upkeep and investments fire.';
-
-  const knowledgeInProgress = Math.max(0, clampNumber(summary.knowledgeInProgress));
-  const knowledgePending = Math.max(0, clampNumber(summary.knowledgePendingToday));
-  const studySummaryText = knowledgeInProgress > 0
-    ? `${knowledgeInProgress} track${knowledgeInProgress === 1 ? '' : 's'} in flight • ${knowledgePending > 0 ? `${knowledgePending} session${knowledgePending === 1 ? '' : 's'} waiting today` : 'All sessions logged today'}`
-    : 'Enroll in a track to kickstart your learning streak.';
-
-  return {
-    time: {
-      summary: timeSummaryText,
-      entries: presentations.timeEntries,
-      emptyMessage: 'Time tracking kicks off after your first action.',
-      limit: 3
-    },
-    earnings: {
-      summary: earningsSummaryText,
-      active: {
-        entries: presentations.earningsEntries,
-        emptyMessage: 'Active gigs will report here.',
-        limit: 3
-      },
-      passive: {
-        entries: presentations.passiveEntries,
-        emptyMessage: 'Passive and offline streams update after upkeep.',
-        limit: 3
-      }
-    },
-    spend: {
-      summary: spendSummaryText,
-      entries: presentations.spendEntries,
-      emptyMessage: 'No cash out yet. Fund upkeep or buy an upgrade.',
-      limit: 3
-    },
-    study: {
-      summary: studySummaryText,
-      entries: presentations.studyEntries,
-      emptyMessage: 'Your courses will list here once enrolled.',
-      limit: 3
-    }
-  };
-}
-
-function buildHeaderMetrics(state = {}, summary = {}) {
-  const hoursLeft = Math.max(0, clampNumber(state.timeLeft));
-  const timeCap = computeTimeCap(state);
-  const dailyEarnings = Math.max(0, clampNumber(summary.totalEarnings));
-  const activeEarnings = Math.max(0, clampNumber(summary.activeEarnings));
-  const passiveEarnings = Math.max(0, clampNumber(summary.passiveEarnings));
-  const dailySpend = Math.max(0, clampNumber(summary.totalSpend));
-  const upkeepSpend = Math.max(0, clampNumber(summary.upkeepSpend));
-  const investmentSpend = Math.max(0, clampNumber(summary.investmentSpend));
-  const reservedHours = Math.max(0, timeCap - hoursLeft);
-  const setupHours = Math.max(0, clampNumber(summary.setupHours));
-  const maintenanceHours = Math.max(0, clampNumber(summary.maintenanceHours));
-  const flexHours = Math.max(0, reservedHours - setupHours - maintenanceHours);
-
-  const incomeSegments = [];
-  if (activeEarnings > 0) {
-    incomeSegments.push(`$${formatMoney(activeEarnings)} active`);
-  }
-  if (passiveEarnings > 0) {
-    incomeSegments.push(`$${formatMoney(passiveEarnings)} passive`);
-  }
-
-  const spendSegments = [];
-  if (upkeepSpend > 0) {
-    spendSegments.push(`$${formatMoney(upkeepSpend)} upkeep`);
-  }
-  if (investmentSpend > 0) {
-    spendSegments.push(`$${formatMoney(investmentSpend)} invest`);
-  }
-  const dailyNet = dailyEarnings - dailySpend;
-  if (dailyNet !== 0) {
-    const netLabel = `${dailyNet >= 0 ? 'Net +' : 'Net -'}$${formatMoney(Math.abs(dailyNet))}`;
-    spendSegments.push(netLabel);
-  }
-
-  const reservedSegments = [];
-  if (setupHours > 0) {
-    reservedSegments.push(`${formatHours(setupHours)} setup`);
-  }
-  if (maintenanceHours > 0) {
-    reservedSegments.push(`${formatHours(maintenanceHours)} upkeep`);
-  }
-  if (flexHours > 0) {
-    reservedSegments.push(`${formatHours(flexHours)} hustle`);
-  }
-
-  return {
-    dailyPlus: {
-      value: `$${formatMoney(dailyEarnings)}`,
-      note: incomeSegments.length ? incomeSegments.join(' • ') : 'Waiting on payouts'
-    },
-    dailyMinus: {
-      value: `$${formatMoney(dailySpend)}`,
-      note: spendSegments.length ? spendSegments.join(' • ') : 'No cash out yet'
-    },
-    timeAvailable: {
-      value: formatHours(hoursLeft),
-      note: `Cap ${formatHours(timeCap)}`
-    },
-    timeReserved: {
-      value: formatHours(reservedHours),
-      note: reservedSegments.length ? reservedSegments.join(' • ') : 'Queue is wide open'
-    }
-  };
-}
-
-function buildKpiStats(state = {}, summary = {}) {
-  const hoursLeft = Math.max(0, clampNumber(state.timeLeft));
-  const net = clampNumber(summary.totalEarnings) - clampNumber(summary.totalSpend);
-  const { activeAssets, upkeepDue } = computeAssetMetrics(state);
-  const { percent, summary: studySummary } = computeStudyProgress(state);
-
-  return {
-    net: {
-      value: `$${formatMoney(net)}`,
-      note: `${formatMoney(clampNumber(summary.totalEarnings))} earned • ${formatMoney(clampNumber(summary.totalSpend))} spent`
-    },
-    hours: {
-      value: `${formatHours(hoursLeft)}`,
-      note: hoursLeft > 0 ? 'Plenty of hustle hours left.' : 'Day is tapped out.'
-    },
-    upkeep: {
-      value: `$${formatMoney(upkeepDue)}`,
-      note: upkeepDue > 0 ? 'Maintain ventures soon.' : 'Upkeep funded.'
-    },
-    ventures: {
-      value: String(activeAssets),
-      note: activeAssets > 0 ? 'Streams humming.' : 'Launch your first venture.'
-    },
-    study: {
-      value: `${percent}%`,
-      note: studySummary
-    }
-  };
-}
-
-function buildQueueModel(summary = {}) {
-  const items = describeQueue(summary).map(entry => ({
-    ...entry,
-    hoursLabel: formatHours(entry.hours)
-  }));
-  return { items };
-}
-
-function buildQuickActionModel(state = {}) {
-  const suggestions = buildQuickActions(state);
-  const entries = suggestions.map(action => ({
-    id: action.id,
-    title: action.label,
-    subtitle: action.description,
-    buttonLabel: action.primaryLabel,
-    onClick: action.onClick,
-    payout: action.payout,
-    payoutText: action.payoutText,
-    durationHours: action.durationHours,
-    durationText: action.durationText,
-    meta: action.meta,
-    repeatable: action.repeatable,
-    remainingRuns: action.remainingRuns
-  }));
-  const baseHours = clampNumber(state.baseTime) + clampNumber(state.bonusTime) + clampNumber(state.dailyBonusTime);
-  const hoursAvailable = Math.max(0, clampNumber(state.timeLeft));
-  const hoursSpent = Math.max(0, baseHours - hoursAvailable);
-  return {
-    entries,
-    emptyMessage: 'No ready actions. Check upgrades or ventures.',
-    buttonClass: 'primary',
-    defaultLabel: 'Queue',
-    hoursAvailable,
-    hoursAvailableLabel: formatHours(hoursAvailable),
-    hoursSpent,
-    hoursSpentLabel: formatHours(hoursSpent),
-    day: clampNumber(state.day),
-    moneyAvailable: clampNumber(state.money)
-  };
-}
-
-function buildAssetActionModel(state = {}) {
-  const suggestions = buildAssetUpgradeRecommendations(state);
-  const entries = suggestions.map(action => {
-    const timeCost = Math.max(0, clampNumber(action.timeCost));
-    return {
-      id: action.id,
-      title: action.title,
-      subtitle: action.subtitle,
-      meta: action.meta,
-      metaClass: 'upgrade-actions__meta',
-      buttonLabel: action.buttonLabel,
-      onClick: action.onClick,
-      timeCost,
-      durationHours: timeCost,
-      durationText: formatHours(timeCost),
-      moneyCost: Math.max(0, clampNumber(action.cost)),
-      repeatable: Boolean(action.repeatable),
-      remainingRuns: action.remainingRuns ?? null
-    };
-  });
-  return {
-    entries,
-    emptyMessage: 'Every venture is humming along. Check back after today’s upkeep.',
-    buttonClass: 'secondary',
-    defaultLabel: 'Boost',
-    scroller: { limit: 6 },
-    moneyAvailable: clampNumber(state.money)
-  };
-}
-
-function buildStudyEnrollmentSuggestions(state = {}) {
-  const suggestions = [];
-
-  if (!state || typeof state !== 'object') {
-    return suggestions;
-  }
-
-  for (const hustle of getHustles()) {
-    if (hustle?.tag?.type !== 'study') continue;
-    const action = hustle?.action;
-    if (!action?.onClick) continue;
-
-    let disabled = false;
-    if (typeof action.disabled === 'function') {
-      try {
-        disabled = action.disabled(state);
-      } catch (error) {
-        disabled = true;
-      }
-    } else {
-      disabled = Boolean(action.disabled);
-    }
-    if (disabled) continue;
-
-    const timeCost = Math.max(0, clampNumber(action.timeCost ?? hustle.time));
-    const tuition = Math.max(0, clampNumber(action.moneyCost));
-    const metaParts = [];
-    if (tuition > 0) {
-      metaParts.push(`$${formatMoney(tuition)} tuition`);
-    }
-    if (timeCost > 0) {
-      metaParts.push(`${formatHours(timeCost)} focus`);
-    }
-
-    const buttonLabel = typeof action.label === 'function'
-      ? action.label(state)
-      : action.label || 'Enroll';
-
-    suggestions.push({
-      id: hustle.id,
-      title: hustle.name,
-      subtitle: hustle.description,
-      meta: metaParts.join(' • '),
-      buttonLabel,
-      onClick: action.onClick,
-      timeCost,
-      durationHours: timeCost,
-      durationText: formatHours(timeCost),
-      moneyCost: tuition,
-      repeatable: false,
-      remainingRuns: null
-    });
-  }
-
-  suggestions.sort((a, b) => a.moneyCost - b.moneyCost);
-  return suggestions;
-}
-
-function buildStudyEnrollmentActionModel(state = {}) {
-  const safeState = state || {};
-  const suggestions = buildStudyEnrollmentSuggestions(safeState);
-  const entries = suggestions.map((action, index) => ({
-    id: action.id || `study-${index}`,
-    title: action.title,
-    subtitle: action.subtitle,
-    meta: action.meta,
-    buttonLabel: action.buttonLabel,
-    onClick: action.onClick,
-    timeCost: action.timeCost,
-    durationHours: action.durationHours,
-    durationText: action.durationText,
-    moneyCost: action.moneyCost,
-    repeatable: action.repeatable,
-    remainingRuns: action.remainingRuns
-  }));
-
-  const baseHours = clampNumber(safeState.baseTime) + clampNumber(safeState.bonusTime) + clampNumber(safeState.dailyBonusTime);
-  const hoursAvailable = Math.max(0, clampNumber(safeState.timeLeft));
-  const hoursSpent = Math.max(0, baseHours - hoursAvailable);
-
-  return {
-    entries,
-    emptyMessage: 'No study tracks are ready to enroll right now.',
-    moneyAvailable: clampNumber(safeState.money),
-    hoursAvailable,
-    hoursAvailableLabel: formatHours(hoursAvailable),
-    hoursSpent,
-    hoursSpentLabel: formatHours(hoursSpent)
-  };
-}
-
 function buildNotificationModel(state = {}) {
   const entries = buildNotifications(state);
   return {
@@ -981,7 +255,7 @@ function buildNotificationModel(state = {}) {
   };
 }
 
-function buildEventLogModel(state = {}) {
+export function buildEventLogModel(state = {}) {
   const allEntries = buildEventLog(state);
   return {
     entries: allEntries.slice(0, 4),
@@ -999,25 +273,26 @@ export function buildDashboardViewModel(state, summary = {}) {
     moneyText: `$${formatMoney(clampNumber(state.money))}`
   };
 
+  const daily = buildDailySummaries(state, summary);
+
   return {
     session,
-    headerMetrics: buildHeaderMetrics(state, summary),
-    kpis: buildKpiStats(state, summary),
-    queue: buildQueueModel(summary),
+    headerMetrics: daily.headerMetrics,
+    kpis: daily.kpis,
+    queue: daily.queue,
     quickActions: buildQuickActionModel(state),
     assetActions: buildAssetActionModel(state),
     studyActions: buildStudyEnrollmentActionModel(state),
     notifications: buildNotificationModel(state),
     eventLog: buildEventLogModel(state),
-    dailyStats: buildDailyStats(summary),
+    dailyStats: daily.dailyStats,
     niche: buildNicheViewModel(state)
   };
 }
 
-export { buildEventLogModel };
-
 export default {
   buildDashboardViewModel,
+  buildDailySummaries,
   buildQuickActions,
   buildAssetUpgradeRecommendations,
   buildStudyEnrollmentActionModel
