+++ src/ui/dashboard/quickActions.js
@@ -1,0 +1,288 @@
+import { formatHours, formatMoney } from '../../core/helpers.js';
+import { clampNumber } from './formatters.js';
+import { getAssetState } from '../../core/state.js';
+import { getAssets, getHustles } from '../../game/registryService.js';
+import {
+  canPerformQualityAction,
+  getNextQualityLevel,
+  getQualityActions,
+  getQualityTracks,
+  performQualityAction
+} from '../../game/assets/quality.js';
+import { instanceLabel } from '../../game/assets/helpers.js';
+
+function getQualitySnapshot(instance = {}) {
+  const level = Math.max(0, clampNumber(instance?.quality?.level));
+  const progress = instance?.quality?.progress && typeof instance.quality.progress === 'object'
+    ? { ...instance.quality.progress }
+    : {};
+  return {
+    level,
+    progress
+  };
+}
+
+function resolveProgressAmount(action, context) {
+  if (!action) return 0;
+  if (typeof action.progressAmount === 'function') {
+    try {
+      const amount = Number(action.progressAmount(context));
+      if (Number.isFinite(amount) && amount > 0) {
+        return amount;
+      }
+    } catch (error) {
+      return 0;
+    }
+  }
+  if (Number.isFinite(Number(action.progressAmount))) {
+    const numeric = Number(action.progressAmount);
+    return numeric > 0 ? numeric : 0;
+  }
+  if (action.progressKey) {
+    return 1;
+  }
+  return 0;
+}
+
+function getProgressPerRun(asset, instance, action, state) {
+  if (!asset || !instance || !action) return 0;
+  const quality = getQualitySnapshot(instance);
+  const context = {
+    state,
+    definition: asset,
+    instance,
+    quality,
+    upgrade: id => state?.upgrades?.[id]
+  };
+  return resolveProgressAmount(action, context);
+}
+
+function estimateRemainingRuns(asset, instance, action, remaining, state) {
+  if (!Number.isFinite(remaining) || remaining <= 0) {
+    return 0;
+  }
+  const progressPerRun = getProgressPerRun(asset, instance, action, state);
+  if (!Number.isFinite(progressPerRun) || progressPerRun <= 0) {
+    return null;
+  }
+  return Math.max(1, Math.ceil(remaining / progressPerRun));
+}
+
+export function buildQuickActions(state) {
+  const items = [];
+  for (const hustle of getHustles()) {
+    if (hustle?.tag?.type === 'study') continue;
+    if (!hustle?.action?.onClick) continue;
+    const usage = typeof hustle.getDailyUsage === 'function' ? hustle.getDailyUsage(state) : null;
+    const remainingRuns = Number.isFinite(usage?.remaining)
+      ? Math.max(0, usage.remaining)
+      : Infinity;
+    const usageLimit = usage?.limit;
+    const repeatable = remainingRuns > 0 && (!Number.isFinite(usageLimit) || usageLimit !== 1);
+    const disabled = typeof hustle.action.disabled === 'function'
+      ? hustle.action.disabled(state)
+      : Boolean(hustle.action.disabled);
+    if (disabled) continue;
+    const payout = clampNumber(hustle.payout?.amount);
+    const time = clampNumber(hustle.time || hustle.action?.timeCost) || 1;
+    const roi = time > 0 ? payout / time : payout;
+    const payoutText = `$${formatMoney(payout)}`;
+    const timeText = formatHours(time);
+    items.push({
+      id: hustle.id,
+      label: hustle.name,
+      primaryLabel: typeof hustle.action.label === 'function'
+        ? hustle.action.label(state)
+        : hustle.action.label || 'Queue',
+      description: `${formatMoney(payout)} payout • ${formatHours(time)}`,
+      onClick: hustle.action.onClick,
+      roi,
+      timeCost: time,
+      payout,
+      payoutText,
+      durationHours: time,
+      durationText: timeText,
+      meta: `${payoutText} • ${timeText}`,
+      repeatable,
+      remainingRuns
+    });
+  }
+
+  items.sort((a, b) => b.roi - a.roi);
+  return items;
+}
+
+export function buildAssetUpgradeRecommendations(state) {
+  if (!state) return [];
+
+  const suggestions = [];
+
+  for (const asset of getAssets()) {
+    const qualityActions = getQualityActions(asset);
+    if (!qualityActions.length) continue;
+
+    const assetState = getAssetState(asset.id, state);
+    const instances = Array.isArray(assetState?.instances) ? assetState.instances : [];
+    if (!instances.length) continue;
+
+    const tracks = getQualityTracks(asset);
+
+    instances.forEach(instance => {
+      if (instance?.status !== 'active') return;
+
+      const quality = instance.quality || {};
+      const level = Math.max(0, clampNumber(quality.level));
+      const nextLevel = getNextQualityLevel(asset, level);
+      if (!nextLevel?.requirements) return;
+
+      const progress = quality.progress || {};
+      const requirements = Object.entries(nextLevel.requirements);
+      if (!requirements.length) return;
+
+      const assetStateInstances = assetState.instances || [];
+      const assetIndex = assetStateInstances.indexOf(instance);
+      const label = instanceLabel(asset, assetIndex >= 0 ? assetIndex : 0);
+      const performance = Math.max(0, clampNumber(instance.lastIncome));
+
+      requirements.forEach(([key, targetValue]) => {
+        const target = Math.max(0, clampNumber(targetValue));
+        if (target <= 0) return;
+        const current = Math.max(0, clampNumber(progress?.[key]));
+        const remaining = Math.max(0, target - current);
+        if (remaining <= 0) return;
+
+        const action = qualityActions.find(entry => entry.progressKey === key);
+        if (!action) return;
+        if (!canPerformQualityAction(asset, instance, action, state)) return;
+
+        const completion = target > 0 ? Math.min(1, current / target) : 1;
+        const percentComplete = Math.max(0, Math.min(100, Math.round(completion * 100)));
+        const percentRemaining = Math.max(0, 100 - percentComplete);
+        const track = tracks?.[key] || {};
+        const requirementLabel = track.shortLabel || track.label || key;
+        const timeCost = Math.max(0, clampNumber(action.time));
+        const moneyCost = Math.max(0, clampNumber(action.cost));
+        const effortParts = [];
+        if (timeCost > 0) {
+          effortParts.push(`${formatHours(timeCost)} focus`);
+        }
+        if (moneyCost > 0) {
+          effortParts.push(`$${formatMoney(moneyCost)}`);
+        }
+        const progressNote = `${Math.min(current, target)}/${target} logged (${percentComplete}% complete)`;
+        const meta = effortParts.length ? `${progressNote} • ${effortParts.join(' • ')}` : progressNote;
+        const actionLabel = typeof action.label === 'function'
+          ? action.label({ definition: asset, instance, state })
+          : action.label;
+        const buttonLabel = actionLabel || 'Boost Quality';
+        const remainingRuns = estimateRemainingRuns(asset, instance, action, remaining, state);
+        const repeatable = remainingRuns == null ? true : remainingRuns > 1;
+
+        suggestions.push({
+          id: `asset-upgrade:${asset.id}:${instance.id}:${action.id}:${key}`,
+          title: `${label} · ${buttonLabel}`,
+          subtitle: `${remaining} ${requirementLabel} to go for Quality ${nextLevel.level} (${percentRemaining}% to go).`,
+          meta,
+          buttonLabel,
+          onClick: () => performQualityAction(asset.id, instance.id, action.id),
+          performance,
+          completion,
+          remaining,
+          level,
+          timeCost,
+          remainingRuns,
+          repeatable,
+          cost: moneyCost
+        });
+      });
+    });
+  }
+
+  suggestions.sort((a, b) => {
+    if (a.performance !== b.performance) {
+      return a.performance - b.performance;
+    }
+    if (a.level !== b.level) {
+      return a.level - b.level;
+    }
+    if (a.completion !== b.completion) {
+      return a.completion - b.completion;
+    }
+    if (a.remaining !== b.remaining) {
+      return b.remaining - a.remaining;
+    }
+    return a.title.localeCompare(b.title);
+  });
+
+  return suggestions;
+}
+
+export function buildQuickActionModel(state = {}) {
+  const suggestions = buildQuickActions(state);
+  const entries = suggestions.map(action => ({
+    id: action.id,
+    title: action.label,
+    subtitle: action.description,
+    buttonLabel: action.primaryLabel,
+    onClick: action.onClick,
+    payout: action.payout,
+    payoutText: action.payoutText,
+    durationHours: action.durationHours,
+    durationText: action.durationText,
+    meta: action.meta,
+    repeatable: action.repeatable,
+    remainingRuns: action.remainingRuns
+  }));
+  const baseHours = clampNumber(state.baseTime) + clampNumber(state.bonusTime) + clampNumber(state.dailyBonusTime);
+  const hoursAvailable = Math.max(0, clampNumber(state.timeLeft));
+  const hoursSpent = Math.max(0, baseHours - hoursAvailable);
+  return {
+    entries,
+    emptyMessage: 'No ready actions. Check upgrades or ventures.',
+    buttonClass: 'primary',
+    defaultLabel: 'Queue',
+    hoursAvailable,
+    hoursAvailableLabel: formatHours(hoursAvailable),
+    hoursSpent,
+    hoursSpentLabel: formatHours(hoursSpent),
+    day: clampNumber(state.day),
+    moneyAvailable: clampNumber(state.money)
+  };
+}
+
+export function buildAssetActionModel(state = {}) {
+  const suggestions = buildAssetUpgradeRecommendations(state);
+  const entries = suggestions.map(action => {
+    const timeCost = Math.max(0, clampNumber(action.timeCost));
+    return {
+      id: action.id,
+      title: action.title,
+      subtitle: action.subtitle,
+      meta: action.meta,
+      metaClass: 'upgrade-actions__meta',
+      buttonLabel: action.buttonLabel,
+      onClick: action.onClick,
+      timeCost,
+      durationHours: timeCost,
+      durationText: formatHours(timeCost),
+      moneyCost: Math.max(0, clampNumber(action.cost)),
+      repeatable: Boolean(action.repeatable),
+      remainingRuns: action.remainingRuns ?? null
+    };
+  });
+  return {
+    entries,
+    emptyMessage: 'Every venture is humming along. Check back after today’s upkeep.',
+    buttonClass: 'secondary',
+    defaultLabel: 'Boost',
+    scroller: { limit: 6 },
+    moneyAvailable: clampNumber(state.money)
+  };
+}
+
+export default {
+  buildQuickActions,
+  buildAssetUpgradeRecommendations,
+  buildQuickActionModel,
+  buildAssetActionModel
+};
